refactor(process): extract process list/kill into windows_mcp.process module
Move list_processes and kill_process out of Desktop service into a new
top-level windows_mcp.process module — they have no dependency on
Desktop state. Update tools/process.py to call the module functions
directly instead of going through get_desktop().
