+++ src/windows_mcp/process/__init__.py
@@ -1,0 +1,6 @@
+from windows_mcp.process.service import (
+    list_processes,
+    kill_process,
+)
+
+__all__ = ["list_processes", "kill_process"]
